Added option to enable duplicates & other changes

+++ instagram.py
@@ -1,4 +1,4 @@
- ######## THIS HORRIBLE (BUT WORKING) CODE WAS MADE BY LANDEN LAFLAMME (TRUFOOX) ########
+ ######## THIS HORRIBLE (BUT WORKING) CODE WAS MADE BY TRUFOOX ########
 import requests, datetime, time, json, sys, math, random, os
 from colorama import Fore, Back, Style
 from PIL import Image
@@ -19,6 +19,7 @@ with open("config.json", "r") as infile: # Starts config
     subredditWeights = config["Subreddit_Weights"]
     blacklist = config["Blacklist"]
     nsfwAllowed = config["NSFW_Allowed"]
+    duplicatesAllowed = config["Duplicates_Allowed"]
     access_token = config["API_Key"]
     user_id = config["User_ID"]
     defaultCaption = config["Use_Post_Caption"]
@@ -40,6 +41,8 @@ publish_url = f"https://graph.facebook.com/v19.0/{user_id}/media_publish"
 tempList = []
 usedDefaultCaption = "False"
 i = 0
+troublesome_subreddits = []
+tSubreddits_times = []
 system("title " + "Instagram Meme Bot") # Sets cmd line window name
 for subreddit in subreddits: # Janky fix to a bug where the subreddit weights arent being utilized
     n = subredditWeights[i]
@@ -92,6 +95,13 @@ while True: # Main code
                 countattempt += 1
                 totalcountattempt += 1
                 response = requests.get(api_url, headers={'User-Agent': 'Mozilla/5.0'})
+                for sub in troublesome_subreddits or []: # 
+                    currentTime = datetime.now().timestamp()
+                    if currentTime - tSubreddits_times[troublesome_subreddits.index(sub)] > timeWait*5:
+                        tSubreddits_times.remove(tSubreddits_times[troublesome_subreddits.index(sub)])
+                        troublesome_subreddits.remove(sub)
+                        print("Just removed troublesome subreddit from list after 5 cycles! lets goooo!!! Remove this line in the code now. Line 102. Also remove 103")
+                        print()
                 if debug == "True":
                     print(f"Response: {response.json()}")
                 if response.status_code == 200: # Test if request sucessful
@@ -101,53 +111,62 @@ while True: # Main code
                         image_url = post_data.get("url", api_url)
                         postCaption = post_data.get("title")
                         print(Fore.WHITE + f"Image URL: {image_url} from r/{chosenSubreddit}. Testing...", end='\r')
-                        if nsfwAllowed == "True" or not nsfw: # Make sure not nsfw if nsfwAllowed is false
-                            if image_shape(image_url):
-                                if not ".gif" in image_url: # Make sure not .gif
-                                    if not image_url in images: # Make sure not duplicate
-                                        if not any(item.lower() in postCaption.lower() for item in blacklist): # Make sure caption doesnt contain blacklisted strings
-                                            if "https://i.redd.it" in image_url: # Prevents weird image URLs (This fixes an error that occurs with non-reddit image URLs)
-                                                images.append(image_url)
-                                                with open("used_images.json", "w") as outfile: # Add successful URL to previously used URLS list
-                                                    json.dump(images, outfile)
-                                                countsuccess += 1
-                                                return image_url
+                        if not chosenSubreddit in troublesome_subreddits:
+                            if nsfwAllowed == "True" or not nsfw: # Makes sure not nsfw if nsfwAllowed is false
+                                if image_shape(image_url):
+                                    if not ".gif" in image_url: # Makes sure not .gif
+                                        if not image_url in images and duplicatesAllowed == "False": # Makes sure not duplicate if duplicates disabled
+                                            if not any(item.lower() in postCaption.lower() for item in blacklist): # Make sure caption doesnt contain blacklisted strings
+                                                if "https://i.redd.it" in image_url: # Prevents weird image URLs (This fixes an error that occurs with non-reddit image URLs)
+                                                    images.append(image_url)
+                                                    with open("used_images.json", "w") as outfile: # Adds successful URL to previously used URLS list
+                                                        json.dump(images, outfile)
+                                                    countsuccess += 1
+                                                    return image_url
+                                                else:
+                                                    delete_previous_line()
+                                                    print(Fore.RED + f"Invalid URL. Trying again... x{countattempt}", flush=True)
                                             else:
                                                 delete_previous_line()
-                                                print(Fore.RED + f"Invalid URL. Trying again... x{countattempt}", flush=True)
+                                                print(Fore.RED + f"Caption contains blacklisted string. Trying again... x{countattempt}", flush=True)
                                         else:
                                             delete_previous_line()
-                                            print(Fore.RED + f"Caption contains blacklisted string. Trying again... x{countattempt}", flush=True)
+                                            print(Fore.RED + f"Duplicate URL. Trying again... x{countattempt}", flush=True)
                                     else:
                                         delete_previous_line()
-                                        print(Fore.RED + f"Duplicate URL. Trying again... x{countattempt}", flush=True)
+                                        print(Fore.RED + f"Image is GIF. Trying again... x{countattempt}", flush=True)
                                 else:
                                     delete_previous_line()
-                                    print(Fore.RED + f"Image is GIF. Trying again... x{countattempt}", flush=True)
+                                    print(Fore.RED + f"Image is not square. Trying again... x{countattempt}", flush=True)
                             else:
                                 delete_previous_line()
-                                print(Fore.RED + f"Image is not square. Trying again... x{countattempt}", flush=True)
+                                print(Fore.RED + f"Image is marked as NSFW. Trying again... x{countattempt}", flush=True)
                         else:
                             delete_previous_line()
-                            print(Fore.RED + f"Image is marked as NSFW. Trying again... x{countattempt}", flush=True)
+                            print(Fore.RED + f"Subreddit has been flagged as overused  Trying again... x{countattempt}", flush=True)
                     else:
-                        print(Fore.RED + f"{datetime.now().strftime('%H:%M')} - Unable to find suitable url in r/{subreddit}. Skipping post. r/{subreddit}", flush=True)
+                        delete_previous_line()
+                        print(Fore.RED + f"{datetime.now().strftime('%H:%M')} - Unable to find suitable url in r/{chosenSubreddit}. Skipping post. Subreddit has been logged as unsuitable for {timeWait*5} minutes", flush=True)
+                        currentTime = datetime.now().timestamp() # Logs subreddit to be ignored for a while (prevents constant loops)
+                        troublesome_subreddits.append(chosenSubreddit)
+                        tSubreddits_times.append(currentTime)
                         print("")
                         return ""
                 else:
                     delete_previous_line()
                     if response.status_code == 530:
-                        print(Fore.RED + style.BOLD + f"{datetime.now().strftime('%H:%M')} - Failed. Cloudfare HTTP Status Code: 530 - The API this program utilizes is inaccessible. Skipping post w/ +6 hour delay." + style.END)
-                        time.sleep(10800)
+                        print(Fore.RED + style.BOLD + f"{datetime.now().strftime('%H:%M')} - Failed. Cloudfare HTTP Status Code: 530 - The API this program utilizes appears to be under maintenence.\nThere is nothing that can be done to fix this but wait. Skipping attempt w/ +6 hour delay..." + style.END)
+                        time.sleep(21600)
                     else:
                         print(Fore.RED + style.BOLD + f"{datetime.now().strftime('%H:%M')} - Failed. HTTP Status Code: {response.status_code} x{countattempt}" + style.END, flush=True)
                         time.sleep(5)
+                    print("")
                 time.sleep(1)
         chosenSubreddit = random.choice(subreddits)
         api_url = f"https://meme-api.com/gimme/{chosenSubreddit}"
         media_url = get_random_image_url(api_url) # Get post from API
-        if not media_url == "": # There was a loop and does not continue if one occurred
-            if defaultCaption == "False" or any(item.lower() in postCaption.lower() for item in captionCriteria): # Set caption to either fallback or post caption
+        if not media_url == "": # Tests if there was a loop/error and does not continue if one occurred
+            if defaultCaption == "False" or any(item.lower() in postCaption.lower() for item in captionCriteria): # Sets caption to either fallback or post caption
                 upload_data = {
                     "image_url": media_url,
                     "caption": caption + '\n\n.' + '\n\n' + hashtags,
@@ -171,7 +190,7 @@ while True: # Main code
                     print(upload_json)
                 if "Error validating access token" in upload_json['error']['message']:
                     print(Fore.RED + f'CRITICAL ERROR: Your Instagram API access token has expired. Follow these instructions to get a new one:')
-                    print(Fore.WHITE + '\n1. Go to this URL to generate an access token (You must have created a Facebook app linked to your Instagram account):' + Fore.CYAN + '\n https://developers.facebook.com/tools/explorer/ ' + Fore.WHITE + '\n\n2. Press the blue "Generate Access Token" button. It will ask you to log in to your facebook account, which is required.\n  - Make sure you log into whichever account owns the Instagram account you intend to use. \n\n3. Go to this url and input the access token you just generated:' + Fore.CYAN + ' \n https://developers.facebook.com/tools/debug/accesstoken' + Fore.WHITE + ' \n\n4. Press the blue "Debug" button. After the new webpage loads, scroll down to the bottom and press "Extend Access Token" \n 5. It will give you a different access token, which will expire in 2 months instead of 1 hour. \n\n 5. Place the result inside the "API_Key" of the config')
+                    print(Fore.WHITE + '\n1. Go to this URL to generate an access token (You must have created a Facebook app linked to your Instagram account):' + Fore.CYAN + '\n https://developers.facebook.com/tools/explorer/ ' + Fore.WHITE + '\n\n2. Press the blue "Generate Access Token" button. It will ask you to log in to your facebook account, which is required.\n  - Make sure you log into whichever account owns the Instagram account you intend to use. \n\n3. Go to this url and input the access token you just generated:' + Fore.CYAN + ' \n https://developers.facebook.com/tools/debug/accesstoken' + Fore.WHITE + ' \n\n4. Press the blue "Debug" button. After the new webpage loads, scroll down to the bottom and press "Extend Access Token" \n5. It will give you a different access token, which will expire in 2 months instead of 1 hour. \n\n6. Place the result inside the "API_Key" of the config')
                     print()
                     input(style.END + "Press Enter to exit...")
                     exit()
@@ -200,7 +219,7 @@ while True: # Main code
     except Exception as error:
         print(Fore.RED + style.BOLD + "Whoops, something is wrong! - Please check your config (See Config Help.txt for help)" + style.END)
         print()
-        print(style.BOLD + "ERROR INFO:" + style.RED, error)
+        print(style.BOLD + "ERROR INFO:" + Foree.RED, error)
         print()
         input(style.END + "Press Enter to exit...")
         exit()
